Initial version of CS162 code.
Students are free to suggest modifications to this code by cloning, modifying and submitting a pull request.  It is suggested that students initially sketch out the modification and confirm with the course instructor that it is a suitable modification.  If the course instructor agrees then go ahead and create the entire modification.  Once merged, then the student will be eligible for a 4 on the most relevant LO for the next assignment.  If 3 such modifications are accepted then the scores may be combined into a 5 (at the students discretion).